Split tests into two structs
